Mudou a note de int pra float

+++ database/migrations/20230523131618_create_table_student_note.js
@@ -10,7 +10,7 @@ export const up = function up(knex) {
             .references("id")
             .inTable("user_discipline");
 
-        table.bigInteger("note")
+        table.float("note")
             .unsigned()
             .notNullable();
     });
